Extract `insertBundles` to the helper
This method builds some applications so that we can use them in various e2e
tests.
